First cut of a data library concept.
A data library is made up of library sections where each sections has a list of data categories. The librarian is in charge of archiving and unarchiving graph data in the library only. The data in a libary section can be filtered by data id, createDate or lastModifiedDate. Also sorting is provided for each data collection by those three data attributes.

The main idea here is the same as aggregate functions in SQL with one big difference: the grouping is done on write events not at runtime. The filter and sort requests are issued against a precomputed data model represented in the heap.